requirements.txt
antlr4-python3-runtime==4.13.1
matplotlib~=3.8.4
numpy~=1.26.4